drop desired codecov level
