+++ codecov.yml
@@ -1,2 +1,6 @@
 ignore:
   - "crispy_code/src/bin/*.rs"
+coverage:
+  precision: 2
+  round: down
+  range: "70...100"
